+++ frontend/src/layout/panel-layout/PanelLayout.tsx
@@ -7,7 +7,7 @@ import { cn } from 'lib/utils/css-classes'
 import { supportTicketCounterLogic } from 'products/conversations/frontend/supportTicketCounterLogic'
 
 import { navigation3000Logic } from '../navigation-3000/navigationLogic'
-import { AiFirstNavBar } from './AiFirstNavBar'
+import { Nav as AiFirstNavBar } from './ai-first/Nav'
 import { panelLayoutLogic } from './panelLayoutLogic'
 import { PanelLayoutNavBar } from './PanelLayoutNavBar'
 import { PROJECT_TREE_KEY, ProjectTree } from './ProjectTree/ProjectTree'
