+++ frontend/src/layout/panel-layout/ai-first/tabs/NavTabChat.tsx
