+++ frontend/src/layout/panel-layout/ai-first/tabs/NavTabInbox.tsx
